Cleanup in nurbs_2d-mocking. Everything is mocked and the long methods are put into a single file which must be ignored by the codefactor checks
