+++ src/spl/weighted_physical_space.h
@@ -16,7 +16,6 @@ You should have received a copy of the GNU Lesser General Public License along w
 #define SRC_SPL_WEIGHTED_PHYSICAL_SPACE_H_
 
 #include <vector>
-#include <iostream>
 
 #include "physical_space.h"
 
@@ -43,7 +42,6 @@ class WeightedPhysicalSpace : public PhysicalSpace<DIM> {
     for (int coordinate = 0; coordinate < this->dimension_; coordinate++) {
       coordinates.push_back(this->control_points_[first + coordinate] * weights_[first / this->dimension_]);
     }
-    std::cout << "GetHomo : indices : " << indices[0] << " , " << indices[1] << ", coordinates " << coordinates[0] << " , " << coordinates[1] << std::endl;
     return baf::ControlPoint(coordinates);
   }
