+++ test/nurbs_2_d_test.cc
@@ -12,233 +12,19 @@ You should have received a copy of the GNU Lesser General Public License along w
 <http://www.gnu.org/licenses/>.
 */
 
-#include <numeric_settings.h>
+#include <array>
+#include <numeric>
+
 #include "gmock/gmock.h"
 
 #include "nurbs.h"
+#include "b_spline.h"
+#include "2d_nurbs_mocking.h"
 
 using testing::Test;
-using testing::DoubleEq;
-using testing::Return;
-using testing::DoubleNear;
 using ::testing::NiceMock;
-using ::testing::Throw;
-using ::testing::_;
-
-class MockParameterSpace1 : public spl::ParameterSpace<2> {
- public:
-  MOCK_CONST_METHOD1(GetDegree, Degree(int));
-  MOCK_CONST_METHOD2(GetBasisFunctions, double(std::array<int, 2>, std::array<ParamCoord, 2>));
-  MOCK_CONST_METHOD3(GetBasisFunctionDerivatives,
-                     double(std::array<int, 2>, std::array<ParamCoord, 2>, std::array<int, 2>));
-  MOCK_CONST_METHOD1(GetArrayOfFirstNonZeroBasisFunctions, std::array<int, 2>(std::array<ParamCoord, 2>));
-  MOCK_CONST_METHOD1(ThrowIfParametricCoordinateOutsideKnotVectorRange, void(std::array<ParamCoord, 2>));
-};
-
-class MockWeightedPhysicalSpace1 : public spl::WeightedPhysicalSpace<2> {
- public:
-  MOCK_CONST_METHOD1(GetWeight, double(std::array<int, 2>));
-  MOCK_CONST_METHOD1(GetHomogenousControlPoint, baf::ControlPoint(std::array<int, 2>));
-  MOCK_CONST_METHOD1(GetControlPoint, baf::ControlPoint(std::array<int, 2>));
-};
-
-void set_get_basis_function_nurbs(const std::shared_ptr<NiceMock<MockParameterSpace1>> &parameter_space) {
-  ON_CALL(*parameter_space, GetBasisFunctions(_, std::array<ParamCoord, 2>{ParamCoord{0.5}, ParamCoord{1.0}}))
-      .WillByDefault(Return(0.0));
-  ON_CALL(*parameter_space, GetBasisFunctions(std::array<int, 2>{0, 2}, std::array<ParamCoord, 2>{ParamCoord{0.5}, ParamCoord{1.0}}))
-      .WillByDefault(Return(0.25));
-  ON_CALL(*parameter_space, GetBasisFunctions(std::array<int, 2>{1, 2}, std::array<ParamCoord, 2>{ParamCoord{0.5}, ParamCoord{1.0}}))
-      .WillByDefault(Return(0.5));
-  ON_CALL(*parameter_space, GetBasisFunctions(std::array<int, 2>{2, 2}, std::array<ParamCoord, 2>{ParamCoord{0.5}, ParamCoord{1.0}}))
-      .WillByDefault(Return(0.25));
-  ON_CALL(*parameter_space, GetBasisFunctions(std::array<int, 2>{0, 0}, std::array<ParamCoord, 2>{ParamCoord{0.4}, ParamCoord{0.6}}))
-      .WillByDefault(Return(0.0576));
-  ON_CALL(*parameter_space, GetBasisFunctions(std::array<int, 2>{1, 0}, std::array<ParamCoord, 2>{ParamCoord{0.4}, ParamCoord{0.6}}))
-      .WillByDefault(Return(0.0768));
-  ON_CALL(*parameter_space, GetBasisFunctions(std::array<int, 2>{2, 0}, std::array<ParamCoord, 2>{ParamCoord{0.4}, ParamCoord{0.6}}))
-      .WillByDefault(Return(0.0256));
-  ON_CALL(*parameter_space, GetBasisFunctions(std::array<int, 2>{0, 1}, std::array<ParamCoord, 2>{ParamCoord{0.4}, ParamCoord{0.6}}))
-      .WillByDefault(Return(0.1728));
-  ON_CALL(*parameter_space, GetBasisFunctions(std::array<int, 2>{1, 1}, std::array<ParamCoord, 2>{ParamCoord{0.4}, ParamCoord{0.6}}))
-      .WillByDefault(Return(0.2304));
-  ON_CALL(*parameter_space, GetBasisFunctions(std::array<int, 2>{2, 1}, std::array<ParamCoord, 2>{ParamCoord{0.4}, ParamCoord{0.6}}))
-      .WillByDefault(Return(0.0768));
-  ON_CALL(*parameter_space, GetBasisFunctions(std::array<int, 2>{0, 2}, std::array<ParamCoord, 2>{ParamCoord{0.4}, ParamCoord{0.6}}))
-      .WillByDefault(Return(0.1296));
-  ON_CALL(*parameter_space, GetBasisFunctions(std::array<int, 2>{1, 2}, std::array<ParamCoord, 2>{ParamCoord{0.4}, ParamCoord{0.6}}))
-      .WillByDefault(Return(0.1728));
-  ON_CALL(*parameter_space, GetBasisFunctions(std::array<int, 2>{2, 2}, std::array<ParamCoord, 2>{ParamCoord{0.4}, ParamCoord{0.6}}))
-      .WillByDefault(Return(0.0576));
-  ON_CALL(*parameter_space, GetBasisFunctions(_, std::array<ParamCoord, 2>{ParamCoord{0.9}, ParamCoord{1.0}}))
-      .WillByDefault(Return(0.0));
-  ON_CALL(*parameter_space, GetBasisFunctions(std::array<int, 2>{0, 2}, std::array<ParamCoord, 2>{ParamCoord{0.9}, ParamCoord{1.0}}))
-      .WillByDefault(Return(0.01));
-  ON_CALL(*parameter_space, GetBasisFunctions(std::array<int, 2>{1, 2}, std::array<ParamCoord, 2>{ParamCoord{0.9}, ParamCoord{1.0}}))
-      .WillByDefault(Return(0.18));
-  ON_CALL(*parameter_space, GetBasisFunctions(std::array<int, 2>{2, 2}, std::array<ParamCoord, 2>{ParamCoord{0.9}, ParamCoord{1.0}}))
-      .WillByDefault(Return(0.81));
-  ON_CALL(*parameter_space, GetBasisFunctions(_, std::array<ParamCoord, 2>{ParamCoord{1.0}, ParamCoord{0.0}}))
-      .WillByDefault(Return(0.0));
-  ON_CALL(*parameter_space, GetBasisFunctions(std::array<int, 2>{2, 0}, std::array<ParamCoord, 2>{ParamCoord{1.0}, ParamCoord{0.0}}))
-      .WillByDefault(Return(1));
-  ON_CALL(*parameter_space, GetBasisFunctions(_, std::array<ParamCoord, 2>{ParamCoord{0.0}, ParamCoord{1.0}}))
-      .WillByDefault(Return(0.0));
-  ON_CALL(*parameter_space, GetBasisFunctions(std::array<int, 2>{0, 2}, std::array<ParamCoord, 2>{ParamCoord{0.0}, ParamCoord{1.0}}))
-      .WillByDefault(Return(1));
-}
-
-void set_basis_function_derivative1(const std::shared_ptr<NiceMock<MockParameterSpace1>> &parameter_space) {
-  ON_CALL(*parameter_space, GetBasisFunctionDerivatives(_, _, _)).WillByDefault(Return(0.0));
-  ON_CALL(*parameter_space, GetBasisFunctionDerivatives(std::array<int, 2>{0, 2},
-                                                        std::array<ParamCoord, 2>{ParamCoord{0.0}, ParamCoord{1.0}},
-                                                        std::array<int, 2>{1, 0}))
-      .WillByDefault(Return(-2.0));
-  ON_CALL(*parameter_space, GetBasisFunctionDerivatives(std::array<int, 2>{1, 2},
-                                                        std::array<ParamCoord, 2>{ParamCoord{0.0}, ParamCoord{1.0}},
-                                                        std::array<int, 2>{1, 0}))
-      .WillByDefault(Return(2.0));
-  ON_CALL(*parameter_space, GetBasisFunctionDerivatives(std::array<int, 2>{0, 1},
-                                                        std::array<ParamCoord, 2>{ParamCoord{0.0}, ParamCoord{1.0}},
-                                                        std::array<int, 2>{0, 1}))
-      .WillByDefault(Return(-2.0));
-  ON_CALL(*parameter_space, GetBasisFunctionDerivatives(std::array<int, 2>{0, 2},
-                                                        std::array<ParamCoord, 2>{ParamCoord{0.0}, ParamCoord{1.0}},
-                                                        std::array<int, 2>{0, 1}))
-      .WillByDefault(Return(2.0));
-  ON_CALL(*parameter_space, GetBasisFunctionDerivatives(std::array<int, 2>{0, 0},
-                                                        std::array<ParamCoord, 2>{ParamCoord{0.4}, ParamCoord{0.6}},
-                                                        std::array<int, 2>{1, 0}))
-      .WillByDefault(Return(-0.192));
-  ON_CALL(*parameter_space, GetBasisFunctionDerivatives(std::array<int, 2>{1, 0},
-                                                        std::array<ParamCoord, 2>{ParamCoord{0.4}, ParamCoord{0.6}},
-                                                        std::array<int, 2>{1, 0}))
-      .WillByDefault(Return(0.064));
-  ON_CALL(*parameter_space, GetBasisFunctionDerivatives(std::array<int, 2>{2, 0},
-                                                        std::array<ParamCoord, 2>{ParamCoord{0.4}, ParamCoord{0.6}},
-                                                        std::array<int, 2>{1, 0}))
-      .WillByDefault(Return(0.128));
-  ON_CALL(*parameter_space, GetBasisFunctionDerivatives(std::array<int, 2>{0, 1},
-                                                        std::array<ParamCoord, 2>{ParamCoord{0.4}, ParamCoord{0.6}},
-                                                        std::array<int, 2>{1, 0}))
-      .WillByDefault(Return(-0.576));
-  ON_CALL(*parameter_space, GetBasisFunctionDerivatives(std::array<int, 2>{1, 1},
-                                                        std::array<ParamCoord, 2>{ParamCoord{0.4}, ParamCoord{0.6}},
-                                                        std::array<int, 2>{1, 0}))
-      .WillByDefault(Return(0.192));
-  ON_CALL(*parameter_space, GetBasisFunctionDerivatives(std::array<int, 2>{2, 1},
-                                                        std::array<ParamCoord, 2>{ParamCoord{0.4}, ParamCoord{0.6}},
-                                                        std::array<int, 2>{1, 0}))
-      .WillByDefault(Return(0.384));
-  ON_CALL(*parameter_space, GetBasisFunctionDerivatives(std::array<int, 2>{0, 2},
-                                                        std::array<ParamCoord, 2>{ParamCoord{0.4}, ParamCoord{0.6}},
-                                                        std::array<int, 2>{1, 0}))
-      .WillByDefault(Return(-0.432));
-  ON_CALL(*parameter_space, GetBasisFunctionDerivatives(std::array<int, 2>{1, 2},
-                                                        std::array<ParamCoord, 2>{ParamCoord{0.4}, ParamCoord{0.6}},
-                                                        std::array<int, 2>{1, 0}))
-      .WillByDefault(Return(0.144));
-  ON_CALL(*parameter_space, GetBasisFunctionDerivatives(std::array<int, 2>{2, 2},
-                                                        std::array<ParamCoord, 2>{ParamCoord{0.4}, ParamCoord{0.6}},
-                                                        std::array<int, 2>{1, 0}))
-      .WillByDefault(Return(0.288));
-  ON_CALL(*parameter_space, GetBasisFunctionDerivatives(std::array<int, 2>{0, 0},
-                                                        std::array<ParamCoord, 2>{ParamCoord{0.4}, ParamCoord{0.6}},
-                                                        std::array<int, 2>{0, 1}))
-      .WillByDefault(Return(-0.288));
-  ON_CALL(*parameter_space, GetBasisFunctionDerivatives(std::array<int, 2>{1, 0},
-                                                        std::array<ParamCoord, 2>{ParamCoord{0.4}, ParamCoord{0.6}},
-                                                        std::array<int, 2>{0, 1}))
-      .WillByDefault(Return(-0.384));
-  ON_CALL(*parameter_space, GetBasisFunctionDerivatives(std::array<int, 2>{2, 0},
-                                                        std::array<ParamCoord, 2>{ParamCoord{0.4}, ParamCoord{0.6}},
-                                                        std::array<int, 2>{0, 1}))
-      .WillByDefault(Return(-0.128));
-  ON_CALL(*parameter_space, GetBasisFunctionDerivatives(std::array<int, 2>{0, 1},
-                                                        std::array<ParamCoord, 2>{ParamCoord{0.4}, ParamCoord{0.6}},
-                                                        std::array<int, 2>{0, 1}))
-      .WillByDefault(Return(-0.144));
-  ON_CALL(*parameter_space, GetBasisFunctionDerivatives(std::array<int, 2>{1, 1},
-                                                        std::array<ParamCoord, 2>{ParamCoord{0.4}, ParamCoord{0.6}},
-                                                        std::array<int, 2>{0, 1}))
-      .WillByDefault(Return(-0.192));
-  ON_CALL(*parameter_space, GetBasisFunctionDerivatives(std::array<int, 2>{2, 1},
-                                                        std::array<ParamCoord, 2>{ParamCoord{0.4}, ParamCoord{0.6}},
-                                                        std::array<int, 2>{0, 1}))
-      .WillByDefault(Return(-0.064));
-  ON_CALL(*parameter_space, GetBasisFunctionDerivatives(std::array<int, 2>{0, 2},
-                                                        std::array<ParamCoord, 2>{ParamCoord{0.4}, ParamCoord{0.6}},
-                                                        std::array<int, 2>{0, 1}))
-      .WillByDefault(Return(0.432));
-  ON_CALL(*parameter_space, GetBasisFunctionDerivatives(std::array<int, 2>{1, 2},
-                                                        std::array<ParamCoord, 2>{ParamCoord{0.4}, ParamCoord{0.6}},
-                                                        std::array<int, 2>{0, 1}))
-      .WillByDefault(Return(0.576));
-  ON_CALL(*parameter_space, GetBasisFunctionDerivatives(std::array<int, 2>{2, 2},
-                                                        std::array<ParamCoord, 2>{ParamCoord{0.4}, ParamCoord{0.6}},
-                                                        std::array<int, 2>{0, 1}))
-      .WillByDefault(Return(0.192));
-}
-
-void mock_parameterSpace_nurbs(const std::shared_ptr<NiceMock<MockParameterSpace1>> &parameter_space) {
-  set_get_basis_function_nurbs(parameter_space);
-  set_basis_function_derivative1(parameter_space);
-  ON_CALL(*parameter_space, GetArrayOfFirstNonZeroBasisFunctions(_))
-      .WillByDefault(Return(std::array<int, 2>{0, 0}));
-  ON_CALL(*parameter_space, GetDegree(_))
-      .WillByDefault(Return(Degree{2}));
-}
-
-void mock_weights(const std::shared_ptr<NiceMock<MockWeightedPhysicalSpace1>> &w_physical_space) {
-  ON_CALL(*w_physical_space, GetWeight(_))
-      .WillByDefault(Return(1));
-  ON_CALL(*w_physical_space, GetWeight(std::array<int, 2>{1, 2}))
-      .WillByDefault(Return(2));
-}
-
-void mock_homogenous(const std::shared_ptr<NiceMock<MockWeightedPhysicalSpace1>> &w_physical_space) {
-  ON_CALL(*w_physical_space, GetHomogenousControlPoint(std::array<int, 2>{0, 0}))
-      .WillByDefault(Return(baf::ControlPoint({0.0, 0.0})));
-  ON_CALL(*w_physical_space, GetHomogenousControlPoint(std::array<int, 2>{1, 0}))
-      .WillByDefault(Return(baf::ControlPoint({1.0, 0.0})));
-  ON_CALL(*w_physical_space, GetHomogenousControlPoint(std::array<int, 2>{2, 0}))
-      .WillByDefault(Return(baf::ControlPoint({3.0, 0.0})));
-  ON_CALL(*w_physical_space, GetHomogenousControlPoint(std::array<int, 2>{0, 1}))
-      .WillByDefault(Return(baf::ControlPoint({-1.0, 0.5})));
-  ON_CALL(*w_physical_space, GetHomogenousControlPoint(std::array<int, 2>{1, 1}))
-      .WillByDefault(Return(baf::ControlPoint({2.0, 2.0})));
-  ON_CALL(*w_physical_space, GetHomogenousControlPoint(std::array<int, 2>{2, 1}))
-      .WillByDefault(Return(baf::ControlPoint({4.0, 1.0})));
-  ON_CALL(*w_physical_space, GetHomogenousControlPoint(std::array<int, 2>{0, 2}))
-      .WillByDefault(Return(baf::ControlPoint({0.0, 2.0})));
-  ON_CALL(*w_physical_space, GetHomogenousControlPoint(std::array<int, 2>{1, 2}))
-      .WillByDefault(Return(baf::ControlPoint({5.0, 7.0})));
-  ON_CALL(*w_physical_space, GetHomogenousControlPoint(std::array<int, 2>{2, 2}))
-      .WillByDefault(Return(baf::ControlPoint({5.0, 2.0})));
-  ON_CALL(*w_physical_space, GetControlPoint(std::array<int, 2>{0, 0}))
-      .WillByDefault(Return(baf::ControlPoint({0.0, 0.0})));
-  ON_CALL(*w_physical_space, GetControlPoint(std::array<int, 2>{1, 0}))
-      .WillByDefault(Return(baf::ControlPoint({1.0, 0.0})));
-  ON_CALL(*w_physical_space, GetControlPoint(std::array<int, 2>{2, 0}))
-      .WillByDefault(Return(baf::ControlPoint({3.0, 0.0})));
-  ON_CALL(*w_physical_space, GetControlPoint(std::array<int, 2>{0, 1}))
-      .WillByDefault(Return(baf::ControlPoint({-1.0, 0.5})));
-  ON_CALL(*w_physical_space, GetControlPoint(std::array<int, 2>{1, 1}))
-      .WillByDefault(Return(baf::ControlPoint({2.0, 2.0})));
-  ON_CALL(*w_physical_space, GetControlPoint(std::array<int, 2>{2, 1}))
-      .WillByDefault(Return(baf::ControlPoint({4.0, 1.0})));
-  ON_CALL(*w_physical_space, GetControlPoint(std::array<int, 2>{0, 2}))
-      .WillByDefault(Return(baf::ControlPoint({0.0, 2.0})));
-  ON_CALL(*w_physical_space, GetControlPoint(std::array<int, 2>{1, 2}))
-      .WillByDefault(Return(baf::ControlPoint({2.5, 3.5})));
-  ON_CALL(*w_physical_space, GetControlPoint(std::array<int, 2>{2, 2}))
-      .WillByDefault(Return(baf::ControlPoint({5.0, 2.0})));
-}
-
-void mock_weightedPhysicalSpace(const std::shared_ptr<NiceMock<MockWeightedPhysicalSpace1>>
-                                &w_physical_space) {
-  mock_weights(w_physical_space);
-  mock_homogenous(w_physical_space);
-}
+using testing::DoubleEq;
+using testing::DoubleNear;
 
 /* 2-dimensional nurbs spline with following properties :
  * KnotVector = {{0, 0, 0, 1, 1, 1}, {0, 0, 0, 1, 1, 1}}
@@ -251,31 +37,6 @@ class A2DNurbs : public Test {
   A2DNurbs() :
       parameter_space(std::make_shared<NiceMock<MockParameterSpace1>>()),
       w_physical_space(std::make_shared<NiceMock<MockWeightedPhysicalSpace1>>()){
-    /*
-    std::array<baf::KnotVector, 2> knot_vector =
-        {baf::KnotVector({ParamCoord{0}, ParamCoord{0}, ParamCoord{0}, ParamCoord{1}, ParamCoord{1}, ParamCoord{1}}),
-         baf::KnotVector({ParamCoord{0}, ParamCoord{0}, ParamCoord{0}, ParamCoord{1}, ParamCoord{1}, ParamCoord{1}})};
-    std::array<Degree, 2> degree = {Degree{2}, Degree{2}};
-    std::vector<double> weights = {1, 1, 1, 1, 1, 1, 1, 2, 1};
-    std::vector<baf::ControlPoint> control_points = {
-        baf::ControlPoint(std::vector<double>({0.0, 0.0})),
-        baf::ControlPoint(std::vector<double>({1.0, 0.0})),
-        baf::ControlPoint(std::vector<double>({3.0, 0.0})),
-        baf::ControlPoint(std::vector<double>({-1.0, 0.5})),
-        baf::ControlPoint(std::vector<double>({2.0, 2.0})),
-        baf::ControlPoint(std::vector<double>({4.0, 1.0})),
-        baf::ControlPoint(std::vector<double>({0.0, 2.0})),
-        baf::ControlPoint(std::vector<double>({2.5, 3.5})),
-        baf::ControlPoint(std::vector<double>({5.0, 2.0}))
-    };
-    std::array<int, 2> number_of_points = {3, 3};
-    std::array<std::shared_ptr<baf::KnotVector>, 2>
-        knot_vector_ptr =
-        {std::make_shared<baf::KnotVector>(knot_vector[0]), std::make_shared<baf::KnotVector>(knot_vector[1])};
-
-    parameter_space = std::make_shared<spl::ParameterSpace<2>>(spl::ParameterSpace<2>(knot_vector_ptr, degree));
-    w_physical_space = std::make_shared<spl::WeightedPhysicalSpace<2>>(spl::WeightedPhysicalSpace<2>(control_points, weights, number_of_points));
-    */
     spl::NURBSGenerator<2> nurbs_generator(w_physical_space, parameter_space);
     nurbs_ = std::make_unique<spl::NURBS<2>>(nurbs_generator);
   }
@@ -383,248 +144,12 @@ TEST_F(A2DNurbs, Returns3_1For0_4And0_6ForDerivative0And1AndDim1) { // NOLINT
               DoubleNear(3.13402, 0.00001));
 }
 
-class MockParameterSpace2 : public spl::ParameterSpace<2> {
- public:
-  MOCK_CONST_METHOD1(GetDegree, Degree(int));
-  MOCK_CONST_METHOD2(GetBasisFunctions, double(std::array<int, 2>, std::array<ParamCoord, 2>));
-  MOCK_CONST_METHOD3(GetBasisFunctionDerivatives,
-                     double(std::array<int, 2>, std::array<ParamCoord, 2>, std::array<int, 2>));
-  MOCK_CONST_METHOD1(GetArrayOfFirstNonZeroBasisFunctions, std::array<int, 2>(std::array<ParamCoord, 2>));
-  MOCK_CONST_METHOD1(ThrowIfParametricCoordinateOutsideKnotVectorRange, void(std::array<ParamCoord, 2>));
-};
 
-class MockPhysicalSpace2 : public spl::PhysicalSpace<2> {
- public:
-  MOCK_CONST_METHOD1(GetControlPoint, baf::ControlPoint(std::array<int, 2>));
- };
-
-class MockWeightedPhysicalSpace2 : public spl::WeightedPhysicalSpace<2> {
- public:
-  MOCK_CONST_METHOD1(GetWeight, double(std::array<int, 2>));
-  MOCK_CONST_METHOD1(GetHomogenousControlPoint, baf::ControlPoint(std::array<int, 2>));
-  MOCK_CONST_METHOD1(GetControlPoint, baf::ControlPoint(std::array<int, 2>));
-};
-
-void set_get_basis_function_nurbs(const std::shared_ptr<NiceMock<MockParameterSpace2>> &parameter_space) {
-  ON_CALL(*parameter_space, GetBasisFunctions(_, std::array<ParamCoord, 2>{ParamCoord{0.0}, ParamCoord{0.7}}))
-      .WillByDefault(Return(0.0));
-  ON_CALL(*parameter_space, GetBasisFunctions(std::array<int, 2>{0, 0}, std::array<ParamCoord, 2>{ParamCoord{0.5}, ParamCoord{0.5}}))
-      .WillByDefault(Return(0.316406));
-  ON_CALL(*parameter_space, GetBasisFunctions(std::array<int, 2>{1, 0}, std::array<ParamCoord, 2>{ParamCoord{0.5}, ParamCoord{0.5}}))
-      .WillByDefault(Return(0.210938));
-  ON_CALL(*parameter_space, GetBasisFunctions(std::array<int, 2>{2, 0}, std::array<ParamCoord, 2>{ParamCoord{0.5}, ParamCoord{0.5}}))
-      .WillByDefault(Return(0.0351562));
-  ON_CALL(*parameter_space, GetBasisFunctions(std::array<int, 2>{0, 1}, std::array<ParamCoord, 2>{ParamCoord{0.5}, ParamCoord{0.5}}))
-      .WillByDefault(Return(0.210938));
-  ON_CALL(*parameter_space, GetBasisFunctions(std::array<int, 2>{1, 1}, std::array<ParamCoord, 2>{ParamCoord{0.5}, ParamCoord{0.5}}))
-      .WillByDefault(Return(0.140625));
-  ON_CALL(*parameter_space, GetBasisFunctions(std::array<int, 2>{2, 1}, std::array<ParamCoord, 2>{ParamCoord{0.5}, ParamCoord{0.5}}))
-      .WillByDefault(Return(0.0234375));
-  ON_CALL(*parameter_space, GetBasisFunctions(std::array<int, 2>{0, 2}, std::array<ParamCoord, 2>{ParamCoord{0.5}, ParamCoord{0.5}}))
-      .WillByDefault(Return(0.0351562));
-  ON_CALL(*parameter_space, GetBasisFunctions(std::array<int, 2>{1, 2}, std::array<ParamCoord, 2>{ParamCoord{0.5}, ParamCoord{0.5}}))
-      .WillByDefault(Return(0.0234375));
-  ON_CALL(*parameter_space, GetBasisFunctions(std::array<int, 2>{2, 2}, std::array<ParamCoord, 2>{ParamCoord{0.5}, ParamCoord{0.5}}))
-      .WillByDefault(Return(0.00390625));
-  ON_CALL(*parameter_space, GetBasisFunctions(std::array<int, 2>{0, 0}, std::array<ParamCoord, 2>{ParamCoord{0.0}, ParamCoord{0.7}}))
-      .WillByDefault(Return(0.4225));
-  ON_CALL(*parameter_space, GetBasisFunctions(std::array<int, 2>{0, 1}, std::array<ParamCoord, 2>{ParamCoord{0.0}, ParamCoord{0.7}}))
-      .WillByDefault(Return(0.455));
-  ON_CALL(*parameter_space, GetBasisFunctions(std::array<int, 2>{0, 2}, std::array<ParamCoord, 2>{ParamCoord{0.0}, ParamCoord{0.7}}))
-      .WillByDefault(Return(0.1225));
-}
-
-void set_basis_function_derivative1(const std::shared_ptr<NiceMock<MockParameterSpace2>> &parameter_space) {
-  //ON_CALL(*parameter_space, GetBasisFunctionDerivatives(_, _, _)).WillByDefault(Return(0.0));
-  ON_CALL(*parameter_space, GetBasisFunctionDerivatives(std::array<int, 2>{0, 0},
-                                                        std::array<ParamCoord, 2>{ParamCoord{0.5}, ParamCoord{0.5}},
-                                                        std::array<int, 2>{1, 1}))
-      .WillByDefault(Return(0.5625));
-  ON_CALL(*parameter_space, GetBasisFunctionDerivatives(std::array<int, 2>{1, 0},
-                                                        std::array<ParamCoord, 2>{ParamCoord{0.5}, ParamCoord{0.5}},
-                                                        std::array<int, 2>{1, 1}))
-      .WillByDefault(Return(-0.375));
-  ON_CALL(*parameter_space, GetBasisFunctionDerivatives(std::array<int, 2>{2, 0},
-                                                        std::array<ParamCoord, 2>{ParamCoord{0.5}, ParamCoord{0.5}},
-                                                        std::array<int, 2>{1, 1}))
-      .WillByDefault(Return(-0.1875));
-  ON_CALL(*parameter_space, GetBasisFunctionDerivatives(std::array<int, 2>{0, 1},
-                                                        std::array<ParamCoord, 2>{ParamCoord{0.5}, ParamCoord{0.5}},
-                                                        std::array<int, 2>{1, 1}))
-      .WillByDefault(Return(-0.375));
-  ON_CALL(*parameter_space, GetBasisFunctionDerivatives(std::array<int, 2>{1, 1},
-                                                        std::array<ParamCoord, 2>{ParamCoord{0.5}, ParamCoord{0.5}},
-                                                        std::array<int, 2>{1, 1}))
-      .WillByDefault(Return(0.25));
-  ON_CALL(*parameter_space, GetBasisFunctionDerivatives(std::array<int, 2>{2, 1},
-                                                        std::array<ParamCoord, 2>{ParamCoord{0.5}, ParamCoord{0.5}},
-                                                        std::array<int, 2>{1, 1}))
-      .WillByDefault(Return(0.125));
-  ON_CALL(*parameter_space, GetBasisFunctionDerivatives(std::array<int, 2>{0, 2},
-                                                        std::array<ParamCoord, 2>{ParamCoord{0.5}, ParamCoord{0.5}},
-                                                        std::array<int, 2>{1, 1}))
-      .WillByDefault(Return(-0.1875));
-  ON_CALL(*parameter_space, GetBasisFunctionDerivatives(std::array<int, 2>{1, 2},
-                                                        std::array<ParamCoord, 2>{ParamCoord{0.5}, ParamCoord{0.5}},
-                                                        std::array<int, 2>{1, 1}))
-      .WillByDefault(Return(0.125));
-  ON_CALL(*parameter_space, GetBasisFunctionDerivatives(std::array<int, 2>{2, 2},
-                                                        std::array<ParamCoord, 2>{ParamCoord{0.5}, ParamCoord{0.5}},
-                                                        std::array<int, 2>{1, 1}))
-      .WillByDefault(Return(0.0625));
-  ON_CALL(*parameter_space, GetBasisFunctionDerivatives(std::array<int, 2>{0, 0},
-                                                        std::array<ParamCoord, 2>{ParamCoord{0.0}, ParamCoord{0.7}},
-                                                        std::array<int, 2>{1, 1}))
-      .WillByDefault(Return(0.65));
-  ON_CALL(*parameter_space, GetBasisFunctionDerivatives(std::array<int, 2>{1, 0},
-                                                        std::array<ParamCoord, 2>{ParamCoord{0.0}, ParamCoord{0.7}},
-                                                        std::array<int, 2>{1, 1}))
-      .WillByDefault(Return(-0.65));
-  ON_CALL(*parameter_space, GetBasisFunctionDerivatives(std::array<int, 2>{2, 0},
-                                                        std::array<ParamCoord, 2>{ParamCoord{0.0}, ParamCoord{0.7}},
-                                                        std::array<int, 2>{1, 1}))
-      .WillByDefault(Return(-0.0));
-  ON_CALL(*parameter_space, GetBasisFunctionDerivatives(std::array<int, 2>{0, 1},
-                                                        std::array<ParamCoord, 2>{ParamCoord{0.0}, ParamCoord{0.7}},
-                                                        std::array<int, 2>{1, 1}))
-      .WillByDefault(Return(-0.3));
-  ON_CALL(*parameter_space, GetBasisFunctionDerivatives(std::array<int, 2>{1, 1},
-                                                        std::array<ParamCoord, 2>{ParamCoord{0.0}, ParamCoord{0.7}},
-                                                        std::array<int, 2>{1, 1}))
-      .WillByDefault(Return(0.3));
-  ON_CALL(*parameter_space, GetBasisFunctionDerivatives(std::array<int, 2>{2, 1},
-                                                        std::array<ParamCoord, 2>{ParamCoord{0.0}, ParamCoord{0.7}},
-                                                        std::array<int, 2>{1, 1}))
-      .WillByDefault(Return(0.0));
-  ON_CALL(*parameter_space, GetBasisFunctionDerivatives(std::array<int, 2>{0, 2},
-                                                        std::array<ParamCoord, 2>{ParamCoord{0.0}, ParamCoord{0.7}},
-                                                        std::array<int, 2>{1, 1}))
-      .WillByDefault(Return(-0.35));
-  ON_CALL(*parameter_space, GetBasisFunctionDerivatives(std::array<int, 2>{1, 2},
-                                                        std::array<ParamCoord, 2>{ParamCoord{0.0}, ParamCoord{0.7}},
-                                                        std::array<int, 2>{1, 1}))
-      .WillByDefault(Return(0.35));
-  ON_CALL(*parameter_space, GetBasisFunctionDerivatives(std::array<int, 2>{2, 2},
-                                                        std::array<ParamCoord, 2>{ParamCoord{0.0}, ParamCoord{0.7}},
-                                                        std::array<int, 2>{1, 1}))
-      .WillByDefault(Return(0.0));
-  ON_CALL(*parameter_space, GetBasisFunctionDerivatives(std::array<int, 2>{0, 0},
-                                                        std::array<ParamCoord, 2>{ParamCoord{0.0}, ParamCoord{0.7}},
-                                                        std::array<int, 2>{2, 1}))
-      .WillByDefault(Return(-0.325));
-  ON_CALL(*parameter_space, GetBasisFunctionDerivatives(std::array<int, 2>{1, 0},
-                                                        std::array<ParamCoord, 2>{ParamCoord{0.0}, ParamCoord{0.7}},
-                                                        std::array<int, 2>{2, 1}))
-      .WillByDefault(Return(0.65));
-  ON_CALL(*parameter_space, GetBasisFunctionDerivatives(std::array<int, 2>{2, 0},
-                                                        std::array<ParamCoord, 2>{ParamCoord{0.0}, ParamCoord{0.7}},
-                                                        std::array<int, 2>{2, 1}))
-      .WillByDefault(Return(-0.325));
-  ON_CALL(*parameter_space, GetBasisFunctionDerivatives(std::array<int, 2>{0, 1},
-                                                        std::array<ParamCoord, 2>{ParamCoord{0.0}, ParamCoord{0.7}},
-                                                        std::array<int, 2>{2, 1}))
-      .WillByDefault(Return(0.15));
-  ON_CALL(*parameter_space, GetBasisFunctionDerivatives(std::array<int, 2>{1, 1},
-                                                        std::array<ParamCoord, 2>{ParamCoord{0.0}, ParamCoord{0.7}},
-                                                        std::array<int, 2>{2, 1}))
-      .WillByDefault(Return(-0.3));
-  ON_CALL(*parameter_space, GetBasisFunctionDerivatives(std::array<int, 2>{2, 1},
-                                                        std::array<ParamCoord, 2>{ParamCoord{0.0}, ParamCoord{0.7}},
-                                                        std::array<int, 2>{2, 1}))
-      .WillByDefault(Return(0.15));
-  ON_CALL(*parameter_space, GetBasisFunctionDerivatives(std::array<int, 2>{0, 2},
-                                                        std::array<ParamCoord, 2>{ParamCoord{0.0}, ParamCoord{0.7}},
-                                                        std::array<int, 2>{2, 1}))
-      .WillByDefault(Return(0.175));
-  ON_CALL(*parameter_space, GetBasisFunctionDerivatives(std::array<int, 2>{1, 2},
-                                                        std::array<ParamCoord, 2>{ParamCoord{0.0}, ParamCoord{0.7}},
-                                                        std::array<int, 2>{2, 1}))
-      .WillByDefault(Return(-0.35));
-  ON_CALL(*parameter_space, GetBasisFunctionDerivatives(std::array<int, 2>{2, 2},
-                                                        std::array<ParamCoord, 2>{ParamCoord{0.0}, ParamCoord{0.7}},
-                                                        std::array<int, 2>{2, 1}))
-      .WillByDefault(Return(0.175));
-}
-
-void mock_parameterSpace_nurbs(const std::shared_ptr<NiceMock<MockParameterSpace2>> &parameter_space) {
-  set_get_basis_function_nurbs(parameter_space);
-  set_basis_function_derivative1(parameter_space);
-  ON_CALL(*parameter_space, GetArrayOfFirstNonZeroBasisFunctions(_))
-      .WillByDefault(Return(std::array<int, 2>{0, 0}));
-  ON_CALL(*parameter_space, GetDegree(_))
-      .WillByDefault(Return(Degree{2}));
-}
-
-void mock_weights(const std::shared_ptr<NiceMock<MockWeightedPhysicalSpace2>> &w_physical_space) {
-  ON_CALL(*w_physical_space, GetWeight(_))
-      .WillByDefault(Return(1));
-}
-
-void mock_homogenous(const std::shared_ptr<NiceMock<MockWeightedPhysicalSpace2>> &w_physical_space) {
-  ON_CALL(*w_physical_space, GetHomogenousControlPoint(std::array<int, 2>{0, 0}))
-      .WillByDefault(Return(baf::ControlPoint({1.0, 2.0})));
-  ON_CALL(*w_physical_space, GetHomogenousControlPoint(std::array<int, 2>{1, 0}))
-      .WillByDefault(Return(baf::ControlPoint({2.0, 2.0})));
-  ON_CALL(*w_physical_space, GetHomogenousControlPoint(std::array<int, 2>{2, 0}))
-      .WillByDefault(Return(baf::ControlPoint({4.0, 2.0})));
-  ON_CALL(*w_physical_space, GetHomogenousControlPoint(std::array<int, 2>{0, 1}))
-      .WillByDefault(Return(baf::ControlPoint({0.0, 2.5})));
-  ON_CALL(*w_physical_space, GetHomogenousControlPoint(std::array<int, 2>{1, 1}))
-      .WillByDefault(Return(baf::ControlPoint({3.0, 4.0})));
-  ON_CALL(*w_physical_space, GetHomogenousControlPoint(std::array<int, 2>{2, 1}))
-      .WillByDefault(Return(baf::ControlPoint({5.0, 3.0})));
-  ON_CALL(*w_physical_space, GetHomogenousControlPoint(std::array<int, 2>{0, 2}))
-      .WillByDefault(Return(baf::ControlPoint({1.0, 4.0})));
-  ON_CALL(*w_physical_space, GetHomogenousControlPoint(std::array<int, 2>{1, 2}))
-      .WillByDefault(Return(baf::ControlPoint({3.5, 5.5})));
-  ON_CALL(*w_physical_space, GetHomogenousControlPoint(std::array<int, 2>{2, 2}))
-      .WillByDefault(Return(baf::ControlPoint({6.0, 4.0})));
-  ON_CALL(*w_physical_space, GetControlPoint(std::array<int, 2>{0, 0}))
-      .WillByDefault(Return(baf::ControlPoint({1.0, 2.0})));
-  ON_CALL(*w_physical_space, GetControlPoint(std::array<int, 2>{1, 0}))
-      .WillByDefault(Return(baf::ControlPoint({2.0, 2.0})));
-  ON_CALL(*w_physical_space, GetControlPoint(std::array<int, 2>{2, 0}))
-      .WillByDefault(Return(baf::ControlPoint({4.0, 2.0})));
-  ON_CALL(*w_physical_space, GetControlPoint(std::array<int, 2>{0, 1}))
-      .WillByDefault(Return(baf::ControlPoint({0.0, 2.5})));
-  ON_CALL(*w_physical_space, GetControlPoint(std::array<int, 2>{1, 1}))
-      .WillByDefault(Return(baf::ControlPoint({3.0, 4.0})));
-  ON_CALL(*w_physical_space, GetControlPoint(std::array<int, 2>{2, 1}))
-      .WillByDefault(Return(baf::ControlPoint({5.0, 3.0})));
-  ON_CALL(*w_physical_space, GetControlPoint(std::array<int, 2>{0, 2}))
-      .WillByDefault(Return(baf::ControlPoint({1.0, 4.0})));
-  ON_CALL(*w_physical_space, GetControlPoint(std::array<int, 2>{1, 2}))
-      .WillByDefault(Return(baf::ControlPoint({3.5, 5.5})));
-  ON_CALL(*w_physical_space, GetControlPoint(std::array<int, 2>{2, 2}))
-      .WillByDefault(Return(baf::ControlPoint({6.0, 4.0})));
-}
-
-void mock_weightedPhysicalSpace(const std::shared_ptr<NiceMock<MockWeightedPhysicalSpace2>>
-                                &w_physical_space) {
-  mock_weights(w_physical_space);
-  mock_homogenous(w_physical_space);
-}
-
-void mock_physicalSpace(const std::shared_ptr<NiceMock<MockPhysicalSpace2>> & physical_space) {
-  ON_CALL(*physical_space, GetControlPoint(std::array<int, 2>{0, 0}))
-      .WillByDefault(Return(baf::ControlPoint({1.0, 2.0})));
-  ON_CALL(*physical_space, GetControlPoint(std::array<int, 2>{1, 0}))
-      .WillByDefault(Return(baf::ControlPoint({2.0, 2.0})));
-  ON_CALL(*physical_space, GetControlPoint(std::array<int, 2>{2, 0}))
-      .WillByDefault(Return(baf::ControlPoint({4.0, 2.0})));
-  ON_CALL(*physical_space, GetControlPoint(std::array<int, 2>{0, 1}))
-      .WillByDefault(Return(baf::ControlPoint({0.0, 2.5})));
-  ON_CALL(*physical_space, GetControlPoint(std::array<int, 2>{1, 1}))
-      .WillByDefault(Return(baf::ControlPoint({3.0, 4.0})));
-  ON_CALL(*physical_space, GetControlPoint(std::array<int, 2>{2, 1}))
-      .WillByDefault(Return(baf::ControlPoint({5.0, 3.0})));
-  ON_CALL(*physical_space, GetControlPoint(std::array<int, 2>{0, 2}))
-      .WillByDefault(Return(baf::ControlPoint({1.0, 4.0})));
-  ON_CALL(*physical_space, GetControlPoint(std::array<int, 2>{1, 2}))
-      .WillByDefault(Return(baf::ControlPoint({3.5, 5.5})));
-  ON_CALL(*physical_space, GetControlPoint(std::array<int, 2>{2, 2}))
-      .WillByDefault(Return(baf::ControlPoint({6.0, 4.0})));
-}
+/* 2-dimensional nurbs spline with following properties :
+ * KnotVector = {{0, 0, 0, 2, 2, 2}, {0, 0, 0, 2, 2, 2}}
+ * ControlPoints = {{1, 2}, {2, 2}, {4, 2}, {0, 2.5}, {3, 4}, {5, 3}, {1, 4}, {3.5, 5.5}, {6, 4}}
+ * Weights = {1, 1, 1, 1, 1, 1, 1, 1, 1}
+*/
 
 class A2DNurbsWithAllWeights1 : public Test {
  public:
@@ -632,51 +157,15 @@ class A2DNurbsWithAllWeights1 : public Test {
       parameter_space_m(std::make_shared<NiceMock<MockParameterSpace2>>()),
       w_physical_space_m(std::make_shared<NiceMock<MockWeightedPhysicalSpace2>>()),
       physical_space_m(std::make_shared<NiceMock<MockPhysicalSpace2>>()) {
-    std::array<baf::KnotVector, 2> knot_vector =
-        {baf::KnotVector({ParamCoord{0}, ParamCoord{0}, ParamCoord{0}, ParamCoord{2}, ParamCoord{2}, ParamCoord{2}}),
-         baf::KnotVector({ParamCoord{0}, ParamCoord{0}, ParamCoord{0}, ParamCoord{2}, ParamCoord{2}, ParamCoord{2}})};
-    std::array<Degree, 2> degree = {Degree{2}, Degree{2}};
-    std::vector<double> weights = {1, 1, 1, 1, 1, 1, 1, 1, 1};
-    std::vector<baf::ControlPoint> control_points = {
-        baf::ControlPoint(std::vector<double>({1.0, 2.0})),
-        baf::ControlPoint(std::vector<double>({2.0, 2.0})),
-        baf::ControlPoint(std::vector<double>({4.0, 2.0})),
-        baf::ControlPoint(std::vector<double>({0.0, 2.5})),
-        baf::ControlPoint(std::vector<double>({3.0, 4.0})),
-        baf::ControlPoint(std::vector<double>({5.0, 3.0})),
-        baf::ControlPoint(std::vector<double>({1.0, 4.0})),
-        baf::ControlPoint(std::vector<double>({3.5, 5.5})),
-        baf::ControlPoint(std::vector<double>({6.0, 4.0}))
-    };
-    std::array<int, 2> number_of_points = {3, 3};
-    std::array<std::shared_ptr<baf::KnotVector>, 2>
-        knot_vector_ptr =
-        {std::make_shared<baf::KnotVector>(knot_vector[0]), std::make_shared<baf::KnotVector>(knot_vector[1])};
-    //nurbs_ = std::make_unique<spl::NURBS<2>>(knot_vector_ptr, degree, control_points, weights);
-    //bspline_ = std::make_unique<spl::BSpline<2>>(knot_vector_ptr, degree, control_points);
-
-    parameter_space = std::make_shared<spl::ParameterSpace<2>>(spl::ParameterSpace<2>(knot_vector_ptr, degree));
-    physical_space = std::make_shared<spl::PhysicalSpace<2>>(spl::PhysicalSpace<2>(control_points, number_of_points));
-    w_physical_space = std::make_shared<spl::WeightedPhysicalSpace<2>>(spl::WeightedPhysicalSpace<2>(control_points, weights, number_of_points));
-    /*
-    spl::NURBSGenerator<2> nurbs_generator(w_physical_space, parameter_space);
-    nurbs_ = std::make_unique<spl::NURBS<2>>(nurbs_generator);
-    spl::BSplineGenerator<2> bspline_generator(physical_space, parameter_space);
-    bspline_ = std::make_unique<spl::BSpline<2>>(bspline_generator);
-    */
     spl::NURBSGenerator<2> nurbs_generator(w_physical_space_m, parameter_space_m);
     nurbs_ = std::make_unique<spl::NURBS<2>>(nurbs_generator);
     spl::BSplineGenerator<2> bspline_generator(physical_space_m, parameter_space_m);
     bspline_ = std::make_unique<spl::BSpline<2>>(bspline_generator);
-    
   }
 
  protected:
   std::unique_ptr<spl::NURBS<2>> nurbs_;
   std::unique_ptr<spl::BSpline<2>> bspline_;
-  std::shared_ptr<spl::ParameterSpace<2>> parameter_space;
-  std::shared_ptr<spl::PhysicalSpace<2>> physical_space;
-  std::shared_ptr<spl::WeightedPhysicalSpace<2>> w_physical_space;
   std::shared_ptr<NiceMock<MockParameterSpace2>> parameter_space_m;
   std::shared_ptr<NiceMock<MockWeightedPhysicalSpace2>>w_physical_space_m;
   std::shared_ptr<NiceMock<MockPhysicalSpace2>> physical_space_m;
